Refactor: move `header_map` to `ckb_shared`
