+++ shared/Cargo.toml
@@ -30,8 +30,11 @@ ckb-channel = { path = "../util/channel", version = "= 0.118.0-pre" }
 ckb-app-config = {path = "../util/app-config", version = "= 0.118.0-pre"}
 ckb-migrate = { path = "../util/migrate", version = "= 0.118.0-pre" }
 once_cell = "1.8.0"
-tempfile.workspace = true
+ckb-util = { path = "../util", version = "= 0.113.0-pre" }
 bitflags = "1.0"
+tokio = { version = "1", features = ["sync"] }
+tempfile.workspace = true
+sled = "0.34.7"
 
 [dev-dependencies]
 ckb-systemtime = { path = "../util/systemtime", version = "= 0.118.0-pre", features = ["enable_faketime"] }
