+++ shared/src/lib.rs
@@ -8,3 +8,6 @@ pub use ckb_snapshot::{Snapshot, SnapshotMgr};
 pub use shared::Shared;
 pub use shared_builder::{SharedBuilder, SharedPackage};
 pub mod block_status;
+pub mod types;
+
+pub use types::header_map::HeaderMap;
