+++ sync/Cargo.toml
@@ -37,7 +37,6 @@ tempfile.workspace = true
 ckb-systemtime = { path = "../util/systemtime", version = "= 0.118.0-pre" }
 dashmap = "4.0"
 keyed_priority_queue = "0.3"
-sled = "0.34.7"
 itertools.workspace = true
 
 [dev-dependencies]
